Set up AccountView action bar for updating transactions
Selecting a transaction from the list view spits it into the details and
shows some buttons for updating or deleting a transaction.
Need to change the data the table view model has to show the date, or if
it was expense, but I'll do that when I actually add the update feature.