Fixed wonky transaction delete
I think it was just that I was getting the outflow bool from the wrong column for the query (accidentally skipped 2).
Meaning they were all false/counted as income when deleted.
Also made a test for deleting expenses, which I should've done earlier.
I noticed that the id for the added transaction didn't go to 4, but the first deleted transaction. Might be a problem in the future.