Started working on budgeting
Wrote test for addCategory, and got it to the point where the assertions fail, rather than compile/sql errors.
Because of that, I created the budget class and added the empty method of addCategory.
I also changed matchtest.cpp to create another budget file to be sure that the I have a control account.