Made the arrows change the budget month shown
Clicking the arrows keeps track of relative month, and change which budgets to show.
They also dissapear when their limit is reached.
Neither the month nor previous/next amount change yet.
I also made the number shown to be the remaining amount.
When the category is clicked I'll have the initial amount be filled in the input field.