Removed update function
It was causing too much frustration to try and retrofit the ability to
update budgets as well.
So the alternative is to make people have to delete then add a new
transaction (I also couldn't get an update method to work with just
doing those functions for the user).
I just commented out the test as a reminder for some of what should be
done by an update function when I get back to it, or for others.