Added shiftMany to essentially reset budgets
monthOne is the date given, then it sets the relative dates, gives the
budget the last set one, and resets the spent amount.
Also added it to shiftBudget and tested it.
This may be premature optimization, but I don't know how slow it would
be for someone to open a file after, say, a year.