Spiffed up BudgetView a bit
There's some visual feedback when hovering over a category to help show
you can click it.
The previous month and next month bits were put in a rectangle with a
mouse area, to allow a wider area for clicking. That too has visual
feedback.
Also fixed an error the category background aligning to the wrong thing.