The sidebar now refreshes the balances when a transaction is added or deleted
Made a function to get the balances in the Sidebar component.
Pass that to AccountView which runs it when a transaction is added or
deleted for on or off budget accounts.
Also had to fix aligning to parent when parent had no alignment for the
account columns (repeater not actual column)